+++ app/src/main/java/com/example/harvardartmuseumsproject/database/GalleryDao.kt
@@ -1,11 +1,7 @@
 package com.example.harvardartmuseumsproject.database
 
-import androidx.paging.PagingSource
 import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.harvardartmuseumsproject.model.Galleries
 import com.example.harvardartmuseumsproject.model.Gallery
 
 @Dao
